backend/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-appconf==1.0.6
django-cors-headers==4.6.0
django-extensions==3.2.3
django-imagekit==5.0.0
django-tinymce==4.1.0
djangorestframework==3.15.2
pilkit==3.0
pillow==11.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.2
whitenoise==6.8.2
